fix: Updated custom processor due to issues
The shared processor was updated to remove the 'subprocess.check_output'
command and instead use 'subprocess.run' as this seemed to provide more
reliable results with the Wiz CLI recipe.

I then removed the unneccesary variables used to point to a interpreter,
this will now only run binary executables, which is fine for my needs.